Refactor chains in config
