+++ packages/config/src/adjustments.ts
@@ -1,10 +1,9 @@
 import { assert } from '@l2beat/shared-pure'
-import { chains } from './chains'
 import { bridges } from './projects/bridges'
 import { layer2s } from './projects/layer2s'
 import { layer3s } from './projects/layer3s'
 import { refactored } from './projects/refactored'
-import type { BaseProject, Bridge, Layer2, Layer3 } from './types'
+import type { BaseProject, Bridge, ChainConfig, Layer2, Layer3 } from './types'
 import { isProjectVerified } from './verification/isVerified'
 
 /**
@@ -18,13 +17,20 @@ import { isProjectVerified } from './verification/isVerified'
  * process.
  */
 export function runConfigAdjustments() {
-  layer2s.forEach(adjustEscrows)
-  layer3s.forEach(adjustEscrows)
-  bridges.forEach(adjustEscrows)
-  refactored.forEach(adjustRefactored)
+  const chains = [...layer2s, ...layer3s, ...bridges, ...refactored]
+    .map((x) => x.chainConfig)
+    .filter((x) => x !== undefined)
+
+  layer2s.forEach((p) => adjustEscrows(p, chains))
+  layer3s.forEach((p) => adjustEscrows(p, chains))
+  bridges.forEach((p) => adjustEscrows(p, chains))
+  refactored.forEach((p) => adjustRefactored(p, chains))
 }
 
-function adjustEscrows(project: Layer2 | Layer3 | Bridge) {
+function adjustEscrows(
+  project: Layer2 | Layer3 | Bridge,
+  chains: ChainConfig[],
+) {
   for (const escrow of project.config.escrows) {
     const chain = chains.find((x) => x.name === escrow.chain)
     assert(chain, `Missing chain: ${escrow.chain}`)
@@ -32,7 +38,7 @@ function adjustEscrows(project: Layer2 | Layer3 | Bridge) {
   }
 }
 
-function adjustRefactored(project: BaseProject) {
+function adjustRefactored(project: BaseProject, chains: ChainConfig[]) {
   if (project.statuses) {
     project.statuses.isUnverified ||= !isProjectVerified(project)
   }
