+++ packages/config/src/index.ts
@@ -16,7 +16,6 @@
  * DO NOT ADD NEW NON-TYPE EXPORTS to @l2beat/config. Thanks!
  */
 
-export { chains } from './chains'
 export { CONTRACTS, HOMEPAGE_MILESTONES, PROJECT_COUNTDOWNS } from './common'
 export { ProjectDiscovery } from './discovery/ProjectDiscovery'
 export {
