+++ packages/config/src/projects/bridges/avalanche.ts
@@ -82,6 +82,19 @@ export const avalanche: Bridge = {
       },
     ],
   },
+  chainConfig: {
+    name: 'avalanche',
+    chainId: 43114,
+    explorerUrl: 'https://snowtrace.io',
+    multicallContracts: [
+      {
+        address: EthereumAddress('0xcA11bde05977b3631167028862bE2a173976CA11'),
+        batchSize: 150,
+        sinceBlock: 11907934,
+        version: '3',
+      },
+    ],
+  },
   riskView: {
     validatedBy: {
       value: 'Third Party',
