+++ packages/config/src/projects/bridges/index.test.ts
@@ -1,10 +1,10 @@
 import { assert, ChainId, UnixTime } from '@l2beat/shared-pure'
 import { expect } from 'earl'
-import { chains } from '../../chains'
 import { NUGGETS } from '../../common'
 import { checkRisk } from '../../test/helpers'
 import { tokenList } from '../../tokens/tokens'
 import type { BridgeTechnology, ProjectTechnologyChoice } from '../../types'
+import { chains } from '../chains'
 import { bridges } from './index'
 
 describe('bridges', () => {
