+++ packages/config/src/projects/layer2s/celo.ts
@@ -1,4 +1,4 @@
-import { UnixTime } from '@l2beat/shared-pure'
+import { EthereumAddress, UnixTime } from '@l2beat/shared-pure'
 import type { Layer2 } from '../../types'
 import { upcomingL2 } from './templates/upcoming'
 
@@ -21,5 +21,25 @@ export const celo: Layer2 = upcomingL2({
       repositories: ['https://github.com/celo-org'],
       socialMedia: ['https://x.com/Celo', 'https://discord.com/invite/celo'],
     },
+  },
+  chainConfig: {
+    name: 'celo',
+    chainId: 42220,
+    explorerUrl: 'https://celoscan.io',
+    explorerApi: {
+      url: 'https://api.celoscan.io/api',
+      type: 'etherscan',
+      missingFeatures: {
+        getContractCreation: true,
+      },
+    },
+    multicallContracts: [
+      {
+        address: EthereumAddress('0xcA11bde05977b3631167028862bE2a173976CA11'),
+        batchSize: 150,
+        sinceBlock: 13112599,
+        version: '3',
+      },
+    ],
   },
 })
