+++ packages/config/src/projects/layer2s/polygonpos.ts
@@ -119,6 +119,24 @@ export const polygonpos: Layer2 = {
       startBlock: 5000000,
     },
   },
+  chainConfig: {
+    name: 'polygonpos',
+    chainId: 137,
+    blockscoutV2ApiUrl: 'https://polygon.blockscout.com/api/v2',
+    explorerUrl: 'https://polygonscan.com',
+    explorerApi: {
+      url: 'https://api.polygonscan.com/api',
+      type: 'etherscan',
+    },
+    multicallContracts: [
+      {
+        address: EthereumAddress('0xcA11bde05977b3631167028862bE2a173976CA11'),
+        batchSize: 150,
+        sinceBlock: 25770160,
+        version: '3',
+      },
+    ],
+  },
   dataAvailability: {
     layer: DA_LAYERS.POLYGON_POS_DA,
     bridge: {
