+++ packages/config/src/projects/layer3s/index.test.ts
@@ -2,9 +2,9 @@ import { expect } from 'earl'
 
 import { assert, ChainId } from '@l2beat/shared-pure'
 import { uniq } from 'lodash'
-import { chains } from '../../chains'
 import { NUGGETS } from '../../common'
 import { tokenList } from '../../tokens/tokens'
+import { chains } from '../chains'
 import { layer2s } from '../layer2s'
 import { layer3s } from './index'
