+++ packages/config/src/tokens/tokens.test.ts
@@ -10,8 +10,8 @@ import {
 } from '@l2beat/shared-pure'
 import { expect } from 'earl'
 import { Contract, providers, utils } from 'ethers'
-import { chains } from '../chains'
 import { bridges } from '../projects/bridges'
+import { chains } from '../projects/chains'
 import { tokenList } from './tokens'
 
 // Github actions sets env as an empty string when secret is not set
